refactor: extract all large pages to separate component files
**Massive simplification** — All pages now under 300-line target

### Pages Simplified

**domains/page.tsx**: 551 → 192 lines (-65%)
- Extracted 7 components to _components.tsx (383 lines)
- _VerificationMethod, _DomainCard, _AddDomainForm, _DomainsEmptyState
- _DomainsList, _DomainsLoadingSkeleton, getVerificationMethods

**conversations/page.tsx**: 468 → 131 lines (-72%)
- Extracted 7 components to _components.tsx (353 lines)
- _ProjectSelector, _LoadingTable, _EmptyProjects, _ConversationsTable
- _ConversationsTableHeader, _ConversationRow, _LoadMoreButton

**orgs/members/page.tsx**: 407 → 109 lines (-73%)
- Extracted 8 components to _components.tsx (310 lines)
- _MembersSkeleton, _PageHeader, _MemberItem, _InvitationItem
- _InviteMemberForm, _MembersList, _PendingInvitationsList
- Export Role type for shared usage

### Impact

**Net code reduction**: -994 lines from main pages (components moved to separate files)
**All pages under 300-line target** ✅

### Files Changed

- 3 main page files: 1017 deletions, 23 insertions
- 3 new _components.tsx files: ~1000 lines extracted

---
